refect code structure and add comment
