+++ backend/src/build-system/handlers/frontend-code-generate/FrontendQueueProcessor.ts
@@ -1,6 +1,18 @@
+/**
+ * FrontendQueueProcessor.ts
+ *
+ * This file defines:
+ * 1) The `FileTask` interface, which describes the basic structure
+ *    of tasks that the queue will handle.
+ * 2) The `CodeTaskQueue` class, a simple FIFO queue for tasks.
+ * 3) The `FrontendQueueProcessor` class, which orchestrates:
+ *    - Writing files
+ *    - Validating builds via `FrontendCodeValidator`
+ *    - Attempting fixes (via LLM) if builds fail
+ */
+
 import { Logger } from '@nestjs/common';
-import { FrontendCodeValidator } from './FrontendCodeValidator';
-import { CodeTaskQueue, FileTask } from './CodeTaskQueue';
+import { FrontendCodeValidator } from './CodeValidator';
 import { readFileSync } from 'fs';
 import { chatSyncWithClocker } from 'src/build-system/utils/handler-helper';
 import { createFileWithRetries } from 'src/build-system/utils/files';
@@ -11,11 +23,49 @@ import {
   generateFileOperationPrompt,
 } from './prompt';
 import { FileOperationManager } from './FileOperationManager';
-import { FixResponseParser } from './FixResponseParser';
 
 import normalizePath from 'normalize-path';
 import path from 'path';
 
+/**
+ * Describes a single file task that the queue will process.
+ *
+ * @property filePath       - The relative path to the file (e.g. "src/components/MyComp.ts").
+ * @property fileContents   - The code content for that file (usually generated by an LLM).
+ * @property dependenciesPath? - Optional: references to any internal dependency file(s).
+ */
+export interface FileTask {
+  filePath: string; // e.g. "src/components/MyComponent.ts"
+  fileContents: string; // the code you got from the LLM
+  dependenciesPath?: string;
+}
+
+export class CodeTaskQueue {
+  private tasks: FileTask[] = [];
+
+  enqueue(task: FileTask) {
+    this.tasks.push(task);
+  }
+
+  dequeue(): FileTask | undefined {
+    return this.tasks.shift();
+  }
+
+  get size(): number {
+    return this.tasks.length;
+  }
+}
+
+/**
+ * FrontendQueueProcessor
+ *
+ * This class orchestrates how each `FileTask` is handled:
+ * 1. Writes the file content to disk.
+ * 2. Runs a build/validation step (via `FrontendCodeValidator`).
+ * 3. If there is a build error, attempts to fix the file by
+ *    invoking an LLM and performing the suggested operations.
+ * 4. Repeats until build succeeds or max attempts are reached.
+ */
 export class FrontendQueueProcessor {
   private logger = new Logger('FrontendQueueProcessor');
 
@@ -122,7 +172,15 @@ export class FrontendQueueProcessor {
   }
 
   /**
-   * Fallback if you have no structured error details.
+   * Fallback fix method if no structured error details are available.
+   * This method interacts with an LLM to propose fixes and then
+   * executes those fixes via FileOperationManager.
+   *
+   * @param currentFullFilePath - The absolute path to the file being fixed.
+   * @param task                - The original FileTask with code and path info.
+   * @param rawErrorText        - The error text from the validator, fed to the LLM.
+   *
+   * @returns The new file path if the file was renamed, or null otherwise.
    */
   private async fixFileGeneric(
     currentFullFilePath: string,
@@ -142,7 +200,6 @@ export class FrontendQueueProcessor {
         this.frontendPath,
         this.renameMap,
       );
-      const parser = new FixResponseParser();
 
       //this.logger.log(fixPrompt);
 
@@ -182,7 +239,7 @@ export class FrontendQueueProcessor {
             },
             {
               role: 'assistant',
-              content: `Let me check my result and I must follow the output format`,
+              content: `Let me check my result and I must follow the output format.`,
             },
           ],
         },
@@ -194,7 +251,10 @@ export class FrontendQueueProcessor {
       this.logger.debug('dependency file Paths ' + task.dependenciesPath);
       const parsed_fixResponse = removeCodeBlockFences(fixResponse);
 
-      let operations = parser.parse(parsed_fixResponse, task.filePath);
+      let operations = fileOperationManager.parse(
+        parsed_fixResponse,
+        task.filePath,
+      );
 
       // **If LLM requested additional files, read them**
       if (operations.some((op) => op.action === 'read')) {
@@ -270,7 +330,10 @@ export class FrontendQueueProcessor {
           'Updated Fix Response with extra context: ' + fixResponse,
         );
         const updated_fixResponse = removeCodeBlockFences(fixResponse);
-        operations = await parser.parse(updated_fixResponse, task.filePath);
+        operations = await fileOperationManager.parse(
+          updated_fixResponse,
+          task.filePath,
+        );
       }
 
       const newFilePath =
